Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/same-origin/meta-referrer/same-origin/http-http/xhr-request/no-redirect/same-origin-insecure.http-expected.txt


PASS The referrer URL is stripped-referrer when a
                                 document served over http requires an http
                                 sub-resource via xhr-request using the meta-referrer
                                 delivery method with no-redirect and when
                                 the target request is same-origin.
